Port upstream 0.55.1: isolate OpenCodex cache state
